* 'main' of https://github.com/Tibame-Group-3/Petradise-SpringBoot: (22 commits)
  news - alter name (remove the word "list") product - add function views - reorganize files (including index, news, pd)
  feat: handle duplicate account
  feat: update sign up ui
  feat: add regex check
  feat: 完成註冊功能
  feat: add form validation
  feat: 地址改為必填
  feat: add dynamic address select
  hotel-修改房型 新增編輯房型頁面
  feat: add taiwan district list
  發現他們刪不得O_O
  product - update mvc
  feat: login success pop back
  feat: sessionStorage -> localStorage
  feat: change href
  feat: login
  hotel- 1.新增房型資訊 2.datatable增加查詢功能
  Add check database count to DatabaseLoader.java
  feat: move "images/" path the other method
  feat: add comment
  ...

+++ src/main/java/tw/idv/petradisespringboot/DatabaseLoader.java
@@ -27,6 +27,9 @@ public class DatabaseLoader {
 
     @Bean
     CommandLineRunner initDatabasePetPic(PetPicRepository petPicRepository) {
+        if (petPicRepository.count() != 0) {
+            return args -> {};
+        }
         petPicRepository.deleteAll();
         return args -> {
             savePetPic(1, 1, petPicRepository);
@@ -41,16 +44,16 @@ public class DatabaseLoader {
     }
 
     private void savePetPic(Integer petId, Integer index, PetPicRepository repo) throws Exception {
-        String imagePath = "images/pet_pic_" + petId + "_" + index + ".jpeg";
+        String imagePath = "pet_pic_" + petId + "_" + index + ".jpeg";
         byte[] image = loadImageBytes(imagePath);
         var pic = new PetPic();
         pic.setPetId(petId);
         pic.setPic(image);
-        logger.info("Preloading " + repo.save(pic));
+        logger.info("Preloading pet pic, pet id: " + repo.save(pic).getPetId());
     }
 
     private byte[] loadImageBytes(String imagePath) throws IOException {
-        Resource resource = loader.getResource("classpath:" + imagePath);
+        Resource resource = loader.getResource("classpath:images/" + imagePath);
         try (InputStream inputStream = resource.getInputStream()) {
             return IOUtils.toByteArray(inputStream);
         }
